- Safe when deleting symlinks of plugin/themes (original is not touched)
- Extract method CroogoTheme::delete()
- First Push from new house :)

+++ Plugin/Extensions/Lib/CroogoPlugin.php
@@ -587,9 +587,16 @@ class CroogoPlugin extends Object {
  *
  * @param string $plugin Plugin name
  * @return boolean true when successful, false or array of error messages when failed
+ * @throws InvalidArgumentException
  */
 	public function delete($plugin) {
+		if (empty($plugin)) {
+			throw new InvalidArgumentException(__d('croogo', 'Invalid plugin'));
+		}
 		$pluginPath = APP . 'Plugin' . DS . $plugin;
+		if (is_link($pluginPath)) {
+			return unlink($pluginPath);
+		}
 		$folder = new Folder();
 		$result = $folder->delete($pluginPath);
 		if ($result !== true) {
